Add bracketed paste functionality into vim
This change allows to paste text into a vim buffer keeping the initial
formatting. It can be useful when copy-pasting text from the internet
into vim. Usually, vim inserts tabs everywhere even those are not
needed. This plugin tackles this problem. So, when pasting vim will now
keep tabs as they are.